separate out disjointset to prevent access to internals
